FDG-6283 refactor section
