+++ src/layouts/explainer/sections/national-debt/national-debt.spec.js
@@ -9,7 +9,8 @@ import nationalDebtSections, {
   nationalDebtDataSources,
   nationalDebtDescriptionGenerator,
   nationalDebtDescriptionAppendix,
-  FundingProgramsSection, DebtTrackingSection, DiveDeeperSection
+  FundingProgramsSection,
+  DebtTrackingSection
 } from "./national-debt";
 import {
   mockExplainerPageResponse,
@@ -122,38 +123,6 @@ describe('Tracking the debt', () => {
     });
     spy.mockClear();
 
-  });
-})
-
-describe('Dive deeper into the debt', () => {
-  it('calls the appropriate analytics event when links are clicked on',  () => {
-    const spy = jest.spyOn(Analytics, 'event');
-    const { getByText } = render(
-      <DiveDeeperSection />
-    );
-
-    const financialReport = getByText('FRUSG_2022.pdf', {exact:false});
-    const americasFiscalFuture = getByText('americas-fiscal-future', {exact:false});
-    const debtCeiling = getByText('whitehouse.gov/cea', {exact:false});
-    const federalBorrowing = getByText('whitehouse.gov/wp-content', {exact:false});
-    const federalNetInterestCost = getByText('cbo.gov/publication/56910', {exact:false});
-    const treasurySecurities = getByText('federalreserve.gov', {exact:false});
-    const reducingDeficit = getByText('cbo.gov/publication/56783', {exact:false});
-    const treasuryBulletin = getByText('treasury-bulletin', {exact:false});
-    const usaSpending = getByText('usaspending.gov', {exact:false});
-
-    const resources = [financialReport, americasFiscalFuture, debtCeiling, federalBorrowing,
-      federalNetInterestCost, treasurySecurities, reducingDeficit, treasuryBulletin, usaSpending];
-
-    resources.forEach(resource => {
-      resource.click();
-      expect(spy).toHaveBeenCalledWith({
-        category: 'Explainers',
-        action: `Citation Click`,
-        label: 'Debt - Dive Deeper into the Debt'
-      });
-      spy.mockClear();
-    })
   });
 })
