+++ src/pages/americas-finance-guide/afg-overview.module.scss.d.ts
@@ -95,11 +95,6 @@ export const growingNationalDebtSectionGraph: string;
 export const footerContainer: string;
 export const lineChartContainer: string;
 export const postGraphContent: string;
-export const diveDeeperQuoteRight: string;
-export const diveDeeperQuote: string;
-export const diveDeeperQuoteLeft: string;
-export const diveDeeperCitation: string;
-export const diveDeeperLink: string;
 export const debtAccordion: string;
 export const debtCeilingAccordion: string;
 export const deficitAccordion: string;
